move clear_folder to utils - call from rail perf manager also to clean /tmp
