+++ src/lamp_py/utils/clear_folder.py
@@ -1,0 +1,17 @@
+import os
+import shutil
+
+
+def clear_folder(folder: str) -> None:
+    """
+    Delete contents of entire folder.
+    """
+    for filename in os.listdir(folder):
+        file_path = os.path.join(folder, filename)
+        try:
+            if os.path.isfile(file_path) or os.path.islink(file_path):
+                os.unlink(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+        except Exception as _:
+            pass
